add anneal-beta option for base surface model
